llm_service/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
pyjwt
supabase
tiktoken
python-multipart
